+++ test/components/blankslate_component_test.rb
@@ -49,10 +49,10 @@ class BlankslateComponentTest < Minitest::Test
     assert_selector(".blankslate-icon[height=32]")
   end
 
-  def test_renders_a_blankslate_component_with_an_icon_with_a_custom_height
+  def test_renders_a_blankslate_component_with_an_icon_with_a_custom_size
     render_inline(Primer::BlankslateComponent.new(
       icon: "octoface",
-      icon_height: 64,
+      icon_size: :large,
       title: "Title"
     ))