+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# main
+
+* Add support for border margins, such as: `border_top: 0`.
+
+    *Natasha Umer*
+
 # 0.0.7
 
 * Use `octicons_helper` v11.0.0.